+++ app/src/main/java/no/hiof/set/g6/app/client/ClientTest.java
@@ -4,9 +4,9 @@ package no.hiof.set.g6.app.client;
 import io.github.heathensoft.jlib.common.text.Ascii;
 import io.github.heathensoft.jlib.lwjgl.window.*;
 import io.netty.channel.ChannelFuture;
-import no.hiof.set.g6.db.net.ClientInstance;
-import no.hiof.set.g6.db.net.JsonPacket;
-import no.hiof.set.g6.db.net.LogEntry;
+import no.hiof.set.g6.net.core.ClientInstance;
+import no.hiof.set.g6.net.core.JsonPacket;
+import no.hiof.set.g6.net.core.LogEntry;
 import org.json.simple.JSONObject;
 import org.lwjgl.glfw.GLFW;
 import org.tinylog.Logger;
