+++ network/src/main/java/no/hiof/set/g6/net/SQLDatabase.java
@@ -1,15 +1,13 @@
-package no.hiof.set.g6.db;
+package no.hiof.set.g6.net;
 
 import no.hiof.set.g6.ny.DatatypeArray;
 import no.hiof.set.g6.ny.LocalUser;
 import no.hiof.set.g6.ny.Locks;
 import no.hiof.set.g6.ny.UserAccount;
 
-import java.nio.file.Path;
+public class SQLDatabase implements PrototypeDB {
 
-public class DemoDatabase implements PrototypeDB {
 
-    private Path path;
 
     @Override
     public LocalUser.Role getUserRole(LocalUser user) {
@@ -49,15 +47,5 @@ public class DemoDatabase implements PrototypeDB {
     @Override
     public boolean editLock(Locks lock) throws Exception {
         return false;
-    }
-
-    @Override
-    public void load() throws Exception {
-        PrototypeDB.super.load();
-    }
-
-    @Override
-    public void save() throws Exception {
-        PrototypeDB.super.save();
     }
 }
