+++ network/src/main/java/no/hiof/set/g6/net/core/AppInterface.java
@@ -1,4 +1,4 @@
-package no.hiof.set.g6.db.net;
+package no.hiof.set.g6.net.core;
 
 
 import io.netty.channel.Channel;
