+++ network/src/main/java/no/hiof/set/g6/net/core/LogEntry.java
@@ -1,4 +1,4 @@
-package no.hiof.set.g6.db.net;
+package no.hiof.set.g6.net.core;
 
 
 /**
