+++ network/src/main/java/no/hiof/set/g6/net/packet/Message.java
@@ -1,4 +1,4 @@
-package no.hiof.set.g6.db.packet;
+package no.hiof.set.g6.net.packet;
 
 
 /**
