+++ settings.gradle.kts
@@ -1,4 +1,4 @@
 
 rootProject.name = "SET_Gruppe_6"
 
-include("app","database","datatype")
+include("app","network","datatype")
